% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ranges_current}
\alias{ranges_current}
\title{Current ranges}
\format{
a raster stack of presence absence (1,0)

For further details see
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/PHYLACINE_1.2.1_Release_notes.pdf}
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/PHYLACINE_1.2_Metadata.pdf}
}
\source{
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2}
}
\usage{
ranges_current
}
\description{
Current ranges
}
\keyword{datasets}
