% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_metadata}
\alias{spatial_metadata}
\title{Spatial metadata}
\format{
Table with spatial metadata

For further details see
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/PHYLACINE_1.2.1_Release_notes.pdf}
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/PHYLACINE_1.2_Metadata.pdf}
}
\source{
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2}
}
\usage{
spatial_metadata
}
\description{
Spatial metadata
}
\keyword{datasets}
